# src/main/openapi/reed-gateway.yaml
openapi: 3.0.1
info:
  description: Gateway API for Reed job search integration. Searches jobs via Reed
    external API.
  title: Jobseeker Copilot - Reed Gateway API
  version: 1.1.0
servers:
- description: Generated server url
  url: http://localhost:8087
...
paths:
...
  /api/reed/search:
    get:
      description: Searches for jobs via Reed external API.
      operationId: searchJobs
      parameters:
      - description: Job title or keyword
        example: Software Engineer
        explode: true
        in: query
        name: query
        required: true
        schema:
          type: string
        style: form
      - description: Location (city or postcode)
        example: London
        explode: true
        in: query
        name: location
        required: false
        schema:
          type: string
        style: form
      - description: Maximum number of results
        example: 10
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReedSearchResponse"
          description: Search completed successfully
        "502":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReedSearchResponse"
          description: Reed API error or rate limit exceeded
      summary: Search jobs on Reed
      tags:
      - Reed Jobs
      x-accepts:
      - '*/*'
components:
  schemas:
...
    ReedJobDto:
      example:
        jobId: jobId
        jobTitle: jobTitle
        employerName: employerName
        locationName: locationName
        minimumSalary: minimumSalary
        maximumSalary: maximumSalary
        currency: currency
        date: date
        jobDescription: jobDescription
        jobUrl: jobUrl
        employmentType: employmentType
      properties:
        jobId:
          type: string
        jobTitle:
          type: string
        employerName:
          type: string
        locationName:
          type: string
        minimumSalary:
          type: string
        maximumSalary:
          type: string
        currency:
          type: string
        date:
          type: string
        jobDescription:
          type: string
        jobUrl:
          type: string
        employmentType:
          type: string
      type: object
    ReedSearchResponse:
      example:
        totalResults: 0
        results:
        - jobId: jobId
          jobTitle: jobTitle
          employerName: employerName
          locationName: locationName
          minimumSalary: minimumSalary
          maximumSalary: maximumSalary
          currency: currency
          date: date
          jobDescription: jobDescription
          jobUrl: jobUrl
          employmentType: employmentType
        - jobId: jobId
          jobTitle: jobTitle
          employerName: employerName
          locationName: locationName
          minimumSalary: minimumSalary
          maximumSalary: maximumSalary
          currency: currency
          date: date
          jobDescription: jobDescription
          jobUrl: jobUrl
          employmentType: employmentType
      properties:
        totalResults:
          format: int32
          type: integer
        results:
          items:
            $ref: "#/components/schemas/ReedJobDto"
          type: array
      type: object